sync_inventory/requirements.txt
ShopifyAPI
fulfil-client
